Change how filter group works
