+++ django_project/frontend/src/redux/actions/filters.jsx
@@ -5,16 +5,14 @@ import {
 
 /**
  * Change state of filter
- * @param {object} groupId ID of filter belong to.
  * @param {object} filterId ID of filter.
  * @param {object} checked Filter data.
  */
-export function changeState(groupId, filterId, checked) {
+export function changeState(filterId, checked) {
   return {
     name: FILTERS_ACTION_NAME,
     type: FILTERS_ACTION_CHANGE_STATUS,
     payload: {
-      groupId: groupId,
       filterId: filterId,
       checked: checked
     }
