+++ django_project/frontend/src/redux/reducers/filters.jsx
@@ -9,9 +9,9 @@ const initialState = []
 export default function filtersReducer(state = initialState, action) {
   switch (action.type) {
     case FILTERS_ACTION_CHANGE_STATUS: {
-      const { groupId, filterId, checked } = action.payload;
+      const { filterId, checked } = action.payload;
       const newState = [...state]
-      newState[groupId].options[filterId].checked = checked
+      newState[filterId].checked = checked
       return newState
     }
     default:
